🏗️ restructure: Improve project organization and modularity
- Created a new `rag_api` directory to keep related parts of the code together.
- Created subdirectories within `rag_api` to group related modules.

This should make the code easier to understand and work with as the project grows!
