+++ rag_api/api/models.py
@@ -1,7 +1,8 @@
 import hashlib
 from enum import Enum
+from typing import List, Optional
+
 from pydantic import BaseModel
-from typing import Optional, List
 
 
 class DocumentResponse(BaseModel):
