+++ rag_api/db/store_factory.py
@@ -1,8 +1,8 @@
 from langchain_community.embeddings import OpenAIEmbeddings
-
-from store import AsyncPgVector, ExtendedPgVector
-from store import AtlasMongoVector
 from pymongo import MongoClient
+
+from rag_api.db.store import AsyncPgVector, AtlasMongoVector, ExtendedPgVector
+
 
 def get_vector_store(
     connection_string: str,
@@ -25,7 +25,9 @@ def get_vector_store(
     elif mode == "atlas-mongo":
         mongo_db = MongoClient(connection_string).get_database()
         mong_collection = mongo_db[collection_name]
-        return AtlasMongoVector(collection=mong_collection, embedding=embeddings, index_name=collection_name)
+        return AtlasMongoVector(
+            collection=mong_collection, embedding=embeddings, index_name=collection_name
+        )
 
     else:
         raise ValueError("Invalid mode specified. Choose 'sync' or 'async'.")
@@ -35,20 +37,25 @@ async def create_index_if_not_exists(conn, table_name: str, column_name: str):
     # Construct index name conventionally
     index_name = f"idx_{table_name}_{column_name}"
     # Check if index exists
-    exists = await conn.fetchval(f"""
+    exists = await conn.fetchval(
+        f"""
         SELECT EXISTS (
             SELECT FROM pg_class c
             JOIN pg_namespace n ON n.oid = c.relnamespace
             WHERE c.relname = $1
             AND n.nspname = 'public'  -- Or specify your schema if different
         );
-    """, index_name)
+    """,
+        index_name,
+    )
     # Create the index if it does not exist
     if not exists:
-        await conn.execute(f"""
+        await conn.execute(
+            f"""
             CREATE INDEX CONCURRENTLY IF NOT EXISTS {index_name}
             ON public.{table_name} ({column_name});
-        """)
+        """
+        )
         print(f"Index {index_name} created on {table_name}.{column_name}")
     else:
         print(f"Index {index_name} already exists on {table_name}.{column_name}")
